Revert "Send Traffic Across all Detected Links for Cluster Validation (#30066)"
This reverts commit 35fcaff37764240935ca3d4f1d451a585ba756f4.
